Material containers is now created
we now have a containers for the material. each material can now have a name and be created empty in the container. we can find material in the container using a name or an ID. when creating a new model, the default material will be the "missing" material (can be changed if wanted)